Remove fake auth method, rename to idle
