+++ app/assets/lcars_splash.png
